Add mode option for moving to root cgroup
Previously it was possible only to disable moving process to root
cgroup. With addition of cgroup v2 things got complicated, especially
when used with systemd logging.

Option -P is changed so it takes arguments on/off/auto. New default is
auto, which first tries to set RR priority and tries to move process
only when setting of priority fails. "on" will always move process to
root cgroup. "off" is just an opposite.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>